+++ Models/Wydatek.cs
@@ -7,5 +7,5 @@ public class Wydatek
     public int IdWydatku { get; set; }
     public int IdPracownika { get; set; }
     public string Cel { get; set; }
-    public decimal Kwota { get; set; }
+    public int Kwota { get; set; }
 }
